pyproject.toml
[project]
name = "langgraph-research-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.0",
    "langchain-community>=0.4.1",
    "langchain-tavily>=0.2.18",
    "langgraph>=1.1.6",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
    "tavily-python>=0.7.23",
]

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.4.24",
]
